modules_allocated
