+++ js/check-string-length.js
@@ -1,0 +1,5 @@
+const checkStringLength = function (stringToCheck, maxLength) {
+  return stringToCheck <= maxLength;
+};
+
+checkStringLength (100, 500);
